Refactor: infrastructure DI を concern ごとのファイルに分割
infrastructure.go を集約ポイントに整理し、各 concern（persistence /
clock / webapi / outbox publisher / security）を独立した submodule
ファイルへ切り出した。httpclient.go の「1 ファイル = 1 module 関数」
規約に揃え、fx 依存グラフをコンポーネント単位で読みやすく保つ。
